+++ types/elsys/elt-2-ultrasonic/uplink.js
@@ -558,17 +558,17 @@ function consume(event) {
       // Warm desk 
       const time = new Date().getTime();
       const state = event.state || {};
-      occupancy.minutesSinceLastOccupancy = 0; // Always give out minutesSinceLastOccupancy for consistancy
+      occupancy.minutesSinceLastOccupied = 0; // Always give out minutesSinceLastOccupied for consistancy
       if (occupancy.occupied) {
         delete state.lastOccupancyTimestamp; // Delete last occupancy timestamp
       } else if (state.lastOccupancyTimestamp !== undefined) {
-        occupancy.minutesSinceLastOccupancy = Math.round((time - state.lastOccupancyTimestamp) / 1000 / 60); // Get free since
+        occupancy.minutesSinceLastOccupied = Math.round((time - state.lastOccupancyTimestamp) / 1000 / 60); // Get free since
       } else if (state.lastOccupancyValue) { //
         state.lastOccupancyTimestamp = time; // Start with first no occupancy
       }
 
-      if (Number.isNaN(occupancy.minutesSinceLastOccupancy)) {
-        occupancy.minutesSinceLastOccupancy = 0;
+      if (Number.isNaN(occupancy.minutesSinceLastOccupied)) {
+        occupancy.minutesSinceLastOccupied = 0;
       }
       state.lastOccupancyValue = occupancy.occupied;