+++ types/elsys/ems-desk/uplink.spec.js
@@ -65,7 +65,7 @@ describe("Elsys EMS Desk uplink", () => {
         assert.equal(value.topic, "occupancy");
         assert.equal(value.data.occupancy, 2);
         assert.equal(value.data.occupied, true);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });