+++ types/elsys/ers-eye/uplink.spec.js
@@ -62,7 +62,7 @@ describe("Elsys eye uplink", () => {
         assert.equal(value.data.motion, 1);
         assert.equal(value.data.occupancy, 1);
         assert.equal(value.data.occupied, true);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });
@@ -116,7 +116,7 @@ describe("Elsys eye uplink", () => {
         assert.equal(value.data.motion, 0);
         assert.equal(value.data.occupancy, 2);
         assert.equal(value.data.occupied, true);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });