+++ types/pnicorp/placePod/uplink.spec.js
@@ -42,7 +42,7 @@ describe("PNIcorp place pod Uplink", () => {
         assert.equal(value.topic, "occupancy");
         assert.equal(value.data.occupancy, 0);
         assert.equal(value.data.occupied, false);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });