+++ types/tektelic/smartRoomPIR/uplink.spec.js
@@ -101,7 +101,7 @@ describe("Tektelic Smart Room PIR Sensor Uplink", () => {
         assert.equal(value.data.motion, 0);
         assert.equal(value.data.occupancy, 0);
         assert.equal(value.data.occupied, false);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });
@@ -132,7 +132,7 @@ describe("Tektelic Smart Room PIR Sensor Uplink", () => {
         assert.equal(value.data.motion, 1);
         assert.equal(value.data.occupancy, 1);
         assert.equal(value.data.occupied, true);
-        assert.equal(value.data.minutesSinceLastOccupancy, 0);
+        assert.equal(value.data.minutesSinceLastOccupied, 0);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });