+++ types/yanzi/presenceMini/uplink.js
@@ -48,17 +48,17 @@ function consume(event) {
 
       // Warm desk 
       const time = new Date().getTime();
-      sample.minutesSinceLastOccupancy = 0; // Always give out minutesSinceLastOccupancy for consistancy
+      sample.minutesSinceLastOccupied = 0; // Always give out minutesSinceLastOccupied for consistancy
       if (sample.occupied) {
         delete state.lastOccupancyTimestamp; // Delete last occupancy timestamp
       } else if (state.lastOccupancyTimestamp !== undefined) {
-        sample.minutesSinceLastOccupancy = Math.round((time - state.lastOccupancyTimestamp) / 1000 / 60); // Get free since
+        sample.minutesSinceLastOccupied = Math.round((time - state.lastOccupancyTimestamp) / 1000 / 60); // Get free since
       } else if (state.lastOccupancyValue) { //
         state.lastOccupancyTimestamp = time; // Start with first no occupancy
       }
 
-      if (Number.isNaN(sample.minutesSinceLastOccupancy)) {
-        sample.minutesSinceLastOccupancy = 0;
+      if (Number.isNaN(sample.minutesSinceLastOccupied)) {
+        sample.minutesSinceLastOccupied = 0;
       }
       state.lastOccupancyValue = sample.occupied;
       emit("state", state);